// app/lib/sdk/validations/comments/comments.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Social API
 * OpenAPI spec version: v1
 */
import * as zod from 'zod';


export const PostApiCommentsBody = zod.object({
  "postId": zod.string().optional(),
  "content": zod.string().optional()
})

export const PostApiCommentsResponse = zod.unknown()

export const PostApiCommentsReplyBody = zod.object({
  "postId": zod.string().optional(),
  "content": zod.string().optional(),
  "parentId": zod.string().optional()
})

export const PostApiCommentsReplyResponse = zod.unknown()

export const PutApiCommentsCommentIdParams = zod.object({
  "commentId": zod.string()
})

export const PutApiCommentsCommentIdBody = zod.object({
  "id": zod.string().optional(),
  "content": zod.string().optional()
})

export const PutApiCommentsCommentIdResponse = zod.unknown()

export const DeleteApiCommentsCommentIdParams = zod.object({
  "commentId": zod.string()
})

export const DeleteApiCommentsCommentIdResponse = zod.unknown()

export const GetApiCommentsCommentIdParams = zod.object({
  "commentId": zod.string()
})

export const GetApiCommentsCommentIdResponse = zod.unknown()

export const GetApiCommentsPostPostIdParams = zod.object({
  "postId": zod.string()
})

export const getApiCommentsPostPostIdQueryPageDefault = 1;
export const getApiCommentsPostPostIdQueryLimitDefault = 10;

export const GetApiCommentsPostPostIdQueryParams = zod.object({
  "page": zod.number().int().default(getApiCommentsPostPostIdQueryPageDefault),
  "limit": zod.number().int().default(getApiCommentsPostPostIdQueryLimitDefault)
})

export const GetApiCommentsPostPostIdResponse = zod.unknown()

export const GetApiCommentsRepliesParentIdParams = zod.object({
  "parentId": zod.string()
})

export const GetApiCommentsRepliesParentIdResponse = zod.unknown()